pyproject.toml
[tool.poetry]
name = "napster-catalog"
version = "0.0.1"
description = ""
authors = ["Gonzalo Sestopal <gonzasestopal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.22.0"
python-dotenv = "^0.10.5"
dotenv-linter = "^0.1.5"
scrapy = "2.5.1"

[tool.poetry.dev-dependencies]
scrapy-new = "^0.2"
black = "^19.10b0"
pylint = "^2.4.4"
pre-commit = "^2.0.1"
isort = "^4.3.21"

[tool.black]
line-length = 99
target-version = ['py38']
include = '\.pyi?$'
exclude = '/(\.eggs|\.git|\.venv|\.scrapy)/'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
